Removed all the errors and updated the icon and the title of website
