+++ frontend/src/App.js
@@ -1,8 +1,8 @@
 import * as React from "react";
-import './App.css';
+import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
-import Home from './pages'
+import Home from "./pages/Home";
 import Branches from "./pages/Branches";
 import Teacher from "./pages/Teacher";
 import Landing from "./pages/Landing";
@@ -10,19 +10,18 @@ import Semester from "./pages/Semester";
 import Note from "./pages/Note";
 import Navbar from "./components/Navbar";
 
-
 export default function App() {
   return (
-    <Routes>
-      <Navbar/>
+    <Router>
+      <Navbar />
       <Routes>
-        <Route exact path="./" exact element={<Home />} />
+        <Route exact path="/" element={<Home />} />
         <Route path="/Branches" element={<Branches />} />
         <Route path="/Landing" element={<Landing />} />
         <Route path="/Note" element={<Note />} />
         <Route path="/Semester" element={<Semester />} />
         <Route path="/Teacher" element={<Teacher />} />
       </Routes>
-    </Routes>
+    </Router>
   );
 }
