--- frontend/src/pages/index.js
@@ -1,12 +1,0 @@
-
-import React from 'react';
- 
-const Home = () => {
-    return (
-        <div>
-            <h1>Welcome to GeeksforGeeks</h1>
-        </div>
-    );
-};
- 
-export default Home;
