--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-router-dom": "^6.15.0"
-  }
-}
